+++ compose/ui/ui/src/webMain/kotlin/androidx/compose/ui/platform/accessibility/ComposeWebSemanticsListener.kt
@@ -16,7 +16,6 @@
 
 package androidx.compose.ui.platform.accessibility
 
-import androidx.collection.MutableIntSet
 import androidx.collection.MutableScatterMap
 import androidx.compose.ui.currentTimeMillis
 import androidx.compose.ui.geometry.Offset
@@ -26,11 +25,8 @@ import androidx.compose.ui.semantics.SemanticsConfiguration
 import androidx.compose.ui.semantics.SemanticsNode
 import androidx.compose.ui.semantics.SemanticsOwner
 import androidx.compose.ui.semantics.SemanticsProperties
-import androidx.compose.ui.semantics.getOrNull
 import androidx.compose.ui.util.fastForEach
 import androidx.compose.ui.util.fastJoinToString
-import kotlin.js.js
-import kotlin.math.abs
 import kotlin.time.Duration.Companion.milliseconds
 import kotlinx.browser.document
 import kotlinx.coroutines.CoroutineScope
@@ -179,12 +175,8 @@ internal class ComposeWebSemanticsListener(
     private val targetParentToChildren = MutableScatterMap<HTMLElement, MutableList<HTMLElement>>()
     private val targetChildToParent = MutableScatterMap<HTMLElement, HTMLElement>()
 
-    // Scroll offsets are in CSS pixels. Pending values come from the current Compose state;
-    // applied values distinguish our own DOM writes from browser- or AT-initiated scrolling.
-    private val appliedScrollOffsets = MutableScatterMap<Int, Offset>()
-    private val pendingScrollOffsets = MutableScatterMap<Int, Offset>()
-    private val scrollSizers = MutableScatterMap<Int, HTMLElement>()
-    private val scrollListenersAttached = MutableIntSet()
+    // Responsible for scroll synchronization between Compose and Dom tree
+    private val a11YScrollState = A11YScrollState(idToA11YNode, a11yNodeToSemanticsNode)
 
     /**
      * Event delegation: Single shared click listener for all a11y nodes with SemanticsActions.OnClick.
@@ -203,41 +195,6 @@ internal class ComposeWebSemanticsListener(
             config[SemanticsActions.OnClick].action?.invoke()
         }
     }
-
-    /**
-     * Translates browser scroll offset changes into Compose semantics scroll actions. Scroll events
-     * do not bubble, so this listener is attached to every A11Y DOM scroll container.
-     */
-    private val onScroll: (Event) -> Unit = onScroll@ { event ->
-        val element = event.target as? HTMLElement ?: return@onScroll
-        val semanticsNode = a11yNodeToSemanticsNode[element] ?: return@onScroll
-        val applied = appliedScrollOffsets[semanticsNode.id] ?: return@onScroll
-        val actual = Offset(element.scrollLeft.toFloat(), element.scrollTop.toFloat())
-        val deltaCssPx = actual - applied
-        if (deltaCssPx.isCloseTo(Offset.Zero)) return@onScroll
-
-        // Subsequent events before the next semantics sync must use the latest DOM offset.
-        appliedScrollOffsets[semanticsNode.id] = actual
-
-        val config = semanticsNode.config
-        if (config.contains(SemanticsProperties.Disabled)) return@onScroll
-
-        val density = semanticsNode.layoutNode.density.density
-        val horizontalDirection =
-            if (config.getOrNull(SemanticsProperties.HorizontalScrollAxisRange)
-                    ?.reverseScrolling == true
-            ) -1f else 1f
-        val verticalDirection =
-            if (config.getOrNull(SemanticsProperties.VerticalScrollAxisRange)
-                    ?.reverseScrolling == true
-            ) -1f else 1f
-
-        config.getOrNull(SemanticsActions.ScrollBy)?.action?.invoke(
-            deltaCssPx.x * density * horizontalDirection,
-            deltaCssPx.y * density * verticalDirection,
-        )
-    }
-
 
     /**
      * Updates the A11Y DOM to mirror all current semantics owners using only necessary structural
@@ -273,16 +230,12 @@ internal class ComposeWebSemanticsListener(
         removedIds.forEach { id ->
             val htmlNode = idToA11YNode.remove(id)
             if (htmlNode != null) {
-                htmlNode.removeEventListener("scroll", onScroll)
                 a11yNodeToSemanticsNode.remove(htmlNode)
             }
-            appliedScrollOffsets.remove(id)
-            pendingScrollOffsets.remove(id)
-            scrollSizers.remove(id)
-            scrollListenersAttached.remove(id)
+            a11YScrollState.onNodeRemoved(id, htmlNode)
         }
 
-        applyScrollOffsets()
+        a11YScrollState.applyScrollOffsets()
         updateInertRoots()
     }
 
@@ -494,7 +447,7 @@ internal class ComposeWebSemanticsListener(
             htmlNode.removeAttribute("aria-modal")
         }
 
-        syncScrollability(semanticsNode, config, htmlNode)
+        a11YScrollState.syncNodeScrollability(semanticsNode, config, htmlNode)
 
         val density = semanticsNode.layoutNode.density.density
         // Use unclipped geometry so descendants retain their content-space positions outside a
@@ -515,9 +468,7 @@ internal class ComposeWebSemanticsListener(
             )
         } else {
             htmlNode.style.position = "absolute"
-            val parentScroll = pendingScrollOffsets[parentSemanticsNode.id]
-                ?: appliedScrollOffsets[parentSemanticsNode.id]
-                ?: Offset.Zero
+            val parentScroll = a11YScrollState.getScrollOffset(parentSemanticsNode)
             setSizeAndPosition(
                 htmlNode,
                 (positionInRoot.x - parentSemanticsNode.positionInRoot.x) / density + parentScroll.x,
@@ -526,113 +477,6 @@ internal class ComposeWebSemanticsListener(
                 height,
             )
         }
-    }
-
-
-    /** Makes scroll semantics visible to the browser as a real CSS scroll container. */
-    private fun syncScrollability(
-        semanticsNode: SemanticsNode,
-        config: SemanticsConfiguration,
-        htmlNode: HTMLElement,
-    ) {
-        val nodeId = semanticsNode.id
-        val verticalRange = config.getOrNull(SemanticsProperties.VerticalScrollAxisRange)
-        val horizontalRange = config.getOrNull(SemanticsProperties.HorizontalScrollAxisRange)
-        val canScroll = (verticalRange != null || horizontalRange != null) &&
-            config.getOrNull(SemanticsActions.ScrollBy)?.action != null
-
-        if (!canScroll) {
-            if (scrollListenersAttached.remove(nodeId)) {
-                htmlNode.removeEventListener("scroll", onScroll)
-                resetScrollContainerStyle(htmlNode)
-                htmlNode.removeAttribute("aria-orientation")
-                scrollSizers.remove(nodeId)?.remove()
-                appliedScrollOffsets.remove(nodeId)
-                pendingScrollOffsets.remove(nodeId)
-            }
-            return
-        }
-
-        setScrollContainerStyle(
-            htmlNode,
-            horizontal = horizontalRange != null,
-            vertical = verticalRange != null,
-        )
-        if ((verticalRange != null) xor (horizontalRange != null)) {
-            htmlNode.setAttribute(
-                "aria-orientation",
-                if (verticalRange != null) "vertical" else "horizontal",
-            )
-        } else {
-            htmlNode.removeAttribute("aria-orientation")
-        }
-
-        val density = semanticsNode.layoutNode.density.density
-        val maxHorizontal = horizontalRange?.maxValue?.invoke()?.coerceAtLeast(0f) ?: 0f
-        val maxVertical = verticalRange?.maxValue?.invoke()?.coerceAtLeast(0f) ?: 0f
-        val viewportWidth = semanticsNode.size.width / density
-        val viewportHeight = semanticsNode.size.height / density
-        val contentWidth =
-            (viewportWidth + maxHorizontal / density).coerceAtMost(MAX_SCROLL_EXTENT_CSS_PX)
-        val contentHeight =
-            (viewportHeight + maxVertical / density).coerceAtMost(MAX_SCROLL_EXTENT_CSS_PX)
-
-        val sizer = scrollSizers.getOrPut(nodeId) { createScrollSizer() }
-        setSizeAndPosition(sizer, 0f, 0f, contentWidth, contentHeight)
-
-        // Lazy layouts expose estimated accessibility offsets rather than physical pixels.
-        // Preserve their current DOM offset instead of interpreting the estimate as scrollTop/Left.
-        val isLazyLayout = config.contains(SemanticsActions.ScrollToIndex)
-        val scrollLeft = if (isLazyLayout) {
-            htmlNode.scrollLeft.toFloat()
-        } else {
-            horizontalRange?.let { range ->
-                val value = range.value().coerceIn(0f, maxHorizontal)
-                ((if (range.reverseScrolling) maxHorizontal - value else value) / density)
-                    .coerceIn(0f, (contentWidth - viewportWidth).coerceAtLeast(0f))
-            } ?: 0f
-        }
-        val scrollTop = if (isLazyLayout) {
-            htmlNode.scrollTop.toFloat()
-        } else {
-            verticalRange?.let { range ->
-                val value = range.value().coerceIn(0f, maxVertical)
-                ((if (range.reverseScrolling) maxVertical - value else value) / density)
-                    .coerceIn(0f, (contentHeight - viewportHeight).coerceAtLeast(0f))
-            } ?: 0f
-        }
-        pendingScrollOffsets[nodeId] = Offset(scrollLeft, scrollTop)
-
-        if (scrollListenersAttached.add(nodeId)) {
-            htmlNode.addEventListener("scroll", onScroll)
-        }
-    }
-
-    /** Applies Compose scroll offsets after DOM structure and scroll extents have been updated. */
-    private fun applyScrollOffsets() {
-        pendingScrollOffsets.forEach { id, offset ->
-            val element = idToA11YNode[id] ?: return@forEach
-            val sizer = scrollSizers[id] ?: return@forEach
-            if (sizer.parentElement !== element || element.firstElementChild !== sizer) {
-                element.insertBefore(sizer, element.firstChild)
-            }
-
-            val actual = Offset(element.scrollLeft.toFloat(), element.scrollTop.toFloat())
-            val lastApplied = appliedScrollOffsets[id]
-            if (lastApplied != null && offset.isCloseTo(lastApplied) && !actual.isCloseTo(lastApplied)) {
-                // Preserve a browser/AT offset until its asynchronous scroll event is handled.
-                return@forEach
-            }
-
-            appliedScrollOffsets[id] = offset
-            if (abs(actual.x - offset.x) >= SCROLL_EPSILON_CSS_PX) {
-                element.scrollLeft = offset.x.toDouble()
-            }
-            if (abs(actual.y - offset.y) >= SCROLL_EPSILON_CSS_PX) {
-                element.scrollTop = offset.y.toDouble()
-            }
-        }
-        pendingScrollOffsets.clear()
     }
 
     /**
@@ -727,7 +571,7 @@ internal class ComposeWebSemanticsListener(
             targetTextAndLinks.add(textParts.last())
         }
 
-        updateTextAndLinks(htmlNode, targetTextAndLinks, scrollSizers[node.id])
+        updateTextAndLinks(htmlNode, targetTextAndLinks, a11YScrollState.getScrollSizer(node))
         deferredChildren?.let { pushNodesForTraversal(it, htmlNode) }
         return htmlNode
     }
@@ -786,9 +630,7 @@ internal class ComposeWebSemanticsListener(
         if (!hasStarted || hasStopped) return
 
         webSemanticsRoot.removeEventListener("click", onClick)
-        scrollListenersAttached.forEach { id ->
-            idToA11YNode[id]?.removeEventListener("scroll", onScroll)
-        }
+        a11YScrollState.clear()
 
         dfsSemanticsNodes.clear()
         dfsA11YParents.clear()
@@ -796,10 +638,6 @@ internal class ComposeWebSemanticsListener(
         a11yNodeToSemanticsNode.clear()
         targetParentToChildren.clear()
         targetChildToParent.clear()
-        appliedScrollOffsets.clear()
-        pendingScrollOffsets.clear()
-        scrollSizers.clear()
-        scrollListenersAttached.clear()
 
         invalidationChannel.close()
         syncTriggerChannel.close()
@@ -808,43 +646,4 @@ internal class ComposeWebSemanticsListener(
         removeAllChildrenOf(webSemanticsRoot)
         hasStopped = true
     }
-}
-
-private const val MAX_SCROLL_EXTENT_CSS_PX = 4_000_000f
-private const val SCROLL_EPSILON_CSS_PX = 0.5f
-
-private fun Offset.isCloseTo(other: Offset): Boolean =
-    abs(x - other.x) < SCROLL_EPSILON_CSS_PX && abs(y - other.y) < SCROLL_EPSILON_CSS_PX
-
-private fun createScrollSizer(): HTMLElement {
-    val sizer = document.createElement("div") as HTMLElement
-    sizer.setAttribute("aria-hidden", "true")
-    sizer.style.position = "absolute"
-    return sizer
-}
-
-private fun setScrollContainerStyle(
-    element: HTMLElement,
-    horizontal: Boolean,
-    vertical: Boolean,
-) {
-    // language=javascript
-    js(
-        """
-        element.style.overflowX = horizontal ? "scroll" : "hidden";
-        element.style.overflowY = vertical ? "scroll" : "hidden";
-        element.style.scrollbarWidth = "none";
-        """
-    )
-}
-
-private fun resetScrollContainerStyle(element: HTMLElement) {
-    // language=javascript
-    js(
-        """
-        element.style.overflowX = "";
-        element.style.overflowY = "";
-        element.style.scrollbarWidth = "";
-        """
-    )
 }
